requirements.txt
appium-python-client==3.1.0  
selenium==4.15.2  
google-generativeai==0.8.5  
python-dotenv==1.0.0  
requests==2.31.0  
Pillow==10.1.0  
opencv-python==4.8.1.78  
numpy==1.24.3  
pytest==7.4.3  
allure-pytest==2.13.2 
